ML Model Service
Wrapping the ML Model as a Service
